requirements.txt
fastapi==0.127.0
uvicorn==0.40.0
sqlalchemy==2.0.45
psycopg2-binary==2.9.11
pydantic==2.12.5
python-dotenv==1.2.1
requests==2.32.5
pandas==2.3.3
scikit-learn==1.8.0
numpy==2.4.0